update and refactor
