+++ thunder/core/jit_ext.py
@@ -780,6 +780,9 @@ def autocast_enter(autocast_obj):
     dtype = unwrap_autocast_obj.fast_dtype
     enabled = unwrap_autocast_obj._enabled
     cache_enabled = unwrap_autocast_obj._cache_enabled
+    # NOTE - We manually map `torch.autocast.__enter__` to `torch.amp.autocast_mode._enter_autocast`
+    #        as it is functional variant of the same and it also shows-up in dynamo provided FX-Graphs
+    #        via thunderFX path.
     thunder_fn = _torch_to_thunder_function_map[torch.amp.autocast_mode._enter_autocast]
     thunder_fn(device, dtype, enabled, cache_enabled)
     return wrap(None, provenance=ProvenanceRecord(PseudoInst.LOOKASIDE, inputs=[autocast_obj.provenance]))
@@ -787,9 +790,11 @@ def autocast_enter(autocast_obj):
 
 @register_general_jit_lookaside(torch.autocast.__exit__)
 def autocast_exit(autocast_obj, exc_type, exc_val, exc_tb):
-    unwrap_autocast_obj = unwrap(autocast_obj)
     thunder_fn = _torch_to_thunder_function_map[torch.amp.autocast_mode._exit_autocast]
     thunder_fn()
+    # NOTE - We manually map `torch.autocast.__exit__` to `torch.amp.autocast_mode._exit_autocast`
+    #        as it is functional variant of the same and it also shows-up in dynamo provided FX-Graphs
+    #        via thunderFX path.
     return wrap(None, provenance=ProvenanceRecord(PseudoInst.LOOKASIDE, inputs=[autocast_obj.provenance]))
