+++ thunder/transforms/autocast.py
@@ -3,7 +3,7 @@ from functools import partial, wraps
 from collections.abc import Callable
 from collections.abc import Sequence
 
-from thunder.core import dtypes, prims
+from thunder.core import dtypes, prims, devices
 from thunder.core.pytree import tree_map, tree_flatten
 from thunder.core.proxies import TensorProxy
 from thunder.core.symbol import BoundSymbolInterface, Symbol
@@ -264,8 +264,6 @@ def disable_autocast():
 
 
 def maybe_apply_autocast(sym):
-    from thunder import _get_cache_info
-
     # NOTE: torch.autocast support
     # In PyTorch eager, enabling autocast allows mixed inputs to operator like `linear`
     # which expect dtypes to be same. This works in PyTorch eager, as dispatcher applies the
@@ -279,31 +277,32 @@ def maybe_apply_autocast(sym):
     # AND we are not under `disable_autocast`
     cd = get_compile_data()
     if cd is None:
+        # This would be `None` with the deprecated `thunder.compile` path
         user_enabled_autocast = False
     else:
         user_enabled_autocast = not cd.autocast_stack.is_empty()
 
     if (
-        (user_enabled_autocast)
+        user_enabled_autocast
         and is_autocast_enabled()
+        # symbol has autocast impl.
         and (autocast_impl := _maybe_get_autocast_rule_for_symbol(sym)) is not None
-    ):  # symbol has autocast impl.
+    ):
 
         @wraps(sym)
         def wrapper(*args, **kwargs):
             # See NOTE: Disable Autocast
             with disable_autocast():
-
-                def is_cpu(p):
+                # Helper to determine which device should be queried for active
+                # autocast context.
+                def is_cpu_tensor(p):
                     if isinstance(p, TensorProxy):
-                        return p.device.type == "cpu"
+                        return p.device.devicetype is devices.DeviceType.CPU
                     return False
 
-                is_cpu = any(tree_flatten(tree_map(is_cpu, (args, kwargs)))[0])
-                thunder_autocast_dtype = cd.autocast_stack.get_dtype_for_device_if_enabled("cpu" if is_cpu else "cuda")
-                if (
-                    thunder_autocast_dtype is None
-                ):  # autocast could be explicitly disabled or enabled for another device.
+                any_cpu = any(tree_flatten(tree_map(is_cpu_tensor, (args, kwargs)))[0])
+                thunder_autocast_dtype = cd.autocast_stack.get_dtype_for_device_if_enabled("cpu" if any_cpu else "cuda")
+                if thunder_autocast_dtype is None:  # We take this path if autocast was enabled for another device.
                     return sym(*args, **kwargs)
                 return partial(autocast_impl, dtype=dtypes.to_dtype(thunder_autocast_dtype))(*args, **kwargs)
